requirements.txt
fastapi
uvicorn
openrouter
python-dotenv
sqlalchemy
pydantic
psycopg2-binary